backend/requirements.txt
Flask==3.0.3
boto3==1.34.162
azure-identity==1.17.1
azure-mgmt-compute==31.0.0
python-dotenv==1.0.1
